Practicing ranges and sizes of variables

+++ homework 1/homework 1.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main()
 {
 	string name, city, country;
-	int age;
+	short int age;
 	float monthly_salary;
 	char gender;
 	bool marital_s;
